+++ __init__.py
@@ -1,0 +1,2 @@
+import logging
+log = logging.getLogger("base")
